apps/agent/requirements.txt
# Vigila.io Local Agent - Dependencias
httpx>=0.25.0
websockets>=12.0
python-dotenv>=1.0.0
netifaces>=0.11.0
wsdiscovery>=2.0.0
onvif-zeep>=0.2.12
psutil>=5.9.0
